youtube-assignment/requirements.txt
streamlit
google-api-python-client
isodate
speechrecognition
pydub
google-genai
google-generativeai
pyaudio